+++ srcs/raycaster/check_if_same_point.c
@@ -6,13 +6,13 @@
 /*   By: jnivala <jnivala@student.hive.fi>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/02/09 14:40:02 by jnivala           #+#    #+#             */
-/*   Updated: 2021/02/19 13:20:20 by jnivala          ###   ########.fr       */
+/*   Updated: 2021/02/19 14:37:30 by jnivala          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../doom_nukem.h"
 
-int				check_if_same_point(double current_angle, t_ray_fov *fov)
+int				check_if_same_point(float current_angle, t_ray_fov *fov)
 {
 
 	if (current_angle == 0.0f
