+++ srcs/raycaster/get_distance.c
@@ -6,13 +6,13 @@
 /*   By: jnivala <jnivala@student.hive.fi>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/02/09 15:22:31 by jnivala           #+#    #+#             */
-/*   Updated: 2021/02/19 13:19:51 by jnivala          ###   ########.fr       */
+/*   Updated: 2021/02/19 14:32:37 by jnivala          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../doom_nukem.h"
 
-double	get_distance(t_xy p1, t_xy p2)
+float	get_distance(t_xy p1, t_xy p2)
 {
 	return (sqrt((p2.x - p1.x) * (p2.x - p1.x) + (p2.y - p1.y) * (p2.y - p1.y)));
 }
