+++ srcs/raycaster/scan_fov.c
@@ -6,25 +6,25 @@
 /*   By: jnivala <jnivala@student.hive.fi>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/02/04 07:59:30 by jnivala           #+#    #+#             */
-/*   Updated: 2021/02/19 14:07:12 by jnivala          ###   ########.fr       */
+/*   Updated: 2021/02/19 14:37:17 by jnivala          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../doom_nukem.h"
 
 
-static double	ft_interpolate_y(t_xy p0, t_xy p1)
+static float	ft_interpolate_y(t_xy p0, t_xy p1)
 {
 	if (p1.x - p0.x == 0)
 		return (p1.y - p0.y);
 	return (p0.y + (0 - p0.x) * ((p1.y - p0.y) / (p1.x - p0.x)));
 }
 
-static double	angle_offset(double current_angle, double screen_offset, double screen_wall)
+static float	angle_offset(float current_angle, float screen_offset, float screen_wall)
 {
-	double		angle_mult_left;
-	double		angle_mult_right;
-	double		euc_offset;
+	float		angle_mult_left;
+	float		angle_mult_right;
+	float		euc_offset;
 
 	euc_offset = screen_offset >= 320 ? screen_offset - 320 : screen_offset;
 	angle_mult_left = FOV - euc_offset * 0.00178373853125;
@@ -32,14 +32,14 @@ static double	angle_offset(double current_angle, double screen_offset, double sc
 	return ((angle_mult_left - angle_mult_right) / screen_wall);
 }
 
-static void		ft_draw_wall(t_xy left, t_xy right, t_frame *frame, double current_angle, int color, t_home *home)
+static void		ft_draw_wall(t_xy left, t_xy right, t_frame *frame, float current_angle, int color, t_home *home)
 {
-	double		screen_wall;
-	double		screen_offset;
-	double		wall_height_left;
-	double		wall_height_right;
-	double		step;
-	double		angle_mult;
+	float		screen_wall;
+	float		screen_offset;
+	float		wall_height_left;
+	float		wall_height_right;
+	float		step;
+	float		angle_mult;
 	int			i;
 
 	i = 0;
@@ -67,10 +67,10 @@ static void		ft_draw_wall(t_xy left, t_xy right, t_frame *frame, double current_
 	}
 }
 
-static double	round_angle(double angle, double *pxl_offset)
+static float	round_angle(float angle, float *pxl_offset)
 {
-	double			angle_as_pixels;
-	static double	temp = 0.0f;
+	float			angle_as_pixels;
+	static float	temp = 0.0f;
 	int				trunc;
 
 	angle_as_pixels = angle / 0.002454369;
@@ -79,10 +79,10 @@ static double	round_angle(double angle, double *pxl_offset)
 	if (*pxl_offset >= 1.0)
 	{
 		*pxl_offset = *pxl_offset - 1.0;
-		return ((double)(trunc + 1.0) * 0.002454369);
+		return ((float)(trunc + 1.0) * 0.002454369);
 	}
 	else
-		return ((double)trunc * 0.002454369);
+		return ((float)trunc * 0.002454369);
 }
 
 int				check_if_same_wall(t_xy a, t_xy b, t_xy right_point)
@@ -103,8 +103,8 @@ void			scan_fov(t_home *home, t_frame *frame)
 	t_ray_fov	fov_left;
 	t_ray_fov	fov_right;
 	t_frame		new_frame;
-	double		current_angle;
-	double		current_pxl;
+	float		current_angle;
+	float		current_pxl;
 
 	current_angle = 0.0;
 	fov_left.left_point = vec2(-1,-1);
