+++ srcs/vec_math/vec2_a.c
@@ -6,13 +6,13 @@
 /*   By: jnivala <jnivala@student.hive.fi>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/11/05 19:15:12 by tmaarela          #+#    #+#             */
-/*   Updated: 2021/02/19 13:16:12 by jnivala          ###   ########.fr       */
+/*   Updated: 2021/02/19 14:37:26 by jnivala          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../doom_nukem.h"
 
-t_xy			vec2(double x, double y)
+t_xy			vec2(float x, float y)
 {
 	return ((t_xy){x, y});
 }
