+++ srcs/vec_math/vec2_b.c
@@ -6,32 +6,32 @@
 /*   By: jnivala <jnivala@student.hive.fi>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/11/05 19:16:10 by tmaarela          #+#    #+#             */
-/*   Updated: 2021/02/19 13:16:38 by jnivala          ###   ########.fr       */
+/*   Updated: 2021/02/19 14:38:26 by jnivala          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../doom_nukem.h"
 
-t_xy			vec2_rot(t_xy v, double angle)
+t_xy			vec2_rot(t_xy v, float angle)
 {
 	return ((t_xy){
-		v.x * cos(angle) - v.y * sin(angle),
-		v.x * sin(angle) + v.y * cos(angle)
+		v.x * cosf(angle) - v.y * sinf(angle),
+		v.x * sinf(angle) + v.y * cosf(angle)
 	});
 }
 
 t_xy			vec2_norm(t_xy v)
 {
-	double mag;
+	float mag;
 
-	mag = sqrt((v.x * v.x) + (v.y * v.y));
+	mag = sqrtf((v.x * v.x) + (v.y * v.y));
 	return ((t_xy){
 		v.x / mag,
 		v.y / mag
 	});
 }
 
-t_xy			vec2_mul(t_xy v, double scalar)
+t_xy			vec2_mul(t_xy v, float scalar)
 {
 	return ((t_xy){
 		v.x * scalar,
@@ -39,22 +39,22 @@ t_xy			vec2_mul(t_xy v, double scalar)
 	});
 }
 
-double			vec2_dot(t_xy a, t_xy b)
+float			vec2_dot(t_xy a, t_xy b)
 {
 	return (a.x * b.x + a.y * b.y);
 }
 
-double			vec2_mag(t_xy a)
+float			vec2_mag(t_xy a)
 {
-	return sqrt(a.x * a.x + a.y * a.y);
+	return sqrtf(a.x * a.x + a.y * a.y);
 }
 
-double			vec2_angle(t_xy a, t_xy b)
+float			vec2_angle(t_xy a, t_xy b)
 {
-	return acos(vec2_dot(a, b) / (vec2_mag(a) * vec2_mag(b)));
+	return acosf(vec2_dot(a, b) / (vec2_mag(a) * vec2_mag(b)));
 }
 
-double			ft_map(double in, t_range from, t_range to)
+float			ft_map(float in, t_range from, t_range to)
 {
 	return (to.x + ((in - from.x) * (to.y - to.x)) / (from.y - from.x));
 }
